Make pathToParser and pathToPsiRoot optional
The two properties of GenerateParserTask are only used when
purgeOldFiles is explicitly set to true, but they still had to be
specified.

As of this commit, if purgeOldFiles is set to true but none of the two
properties are specified, the task will delete the whole content of
targetRootOutputDir.

I am not aware of any reason not to delete the whole directory of
targetRootOutputDir unless the directory is overlapping with other files
than the files generated by this task. Since overlapping task outputs
are explicitly discouraged by Gradle, I deprecated the two properties.